Update svelte-kit.ts
update to the sveltekit handler. It's intended to attach the user session to events.locals so that it's accessible through out the app. 

This will give access to the session (and user data) in Page / Layout load functions [ with event.locals.session | event.locals.user ] and Form actions [ with event.locals.session | event.locals.user ]. 

For Svelte files, we can grab the session in a load function and pass it down to the svelte page / layout file.

For TS: requires updating Locals interface in app.d.ts by adding session and user properties with type Session and User [ from "better-auth/types" ].