Remix documentation: Change file name from ``auth.client.ts`` to ``auth-client.ts``
``auth.client.ts`` will trigger code splitting. While this would be fine for the server component (``auth.server.ts``, which needs to stay hidden to end users), putting the authClient in a client only component will force the authClient out of the server bundle. This causes authClient to be undefined during SSR.