docs(error-handling): added error message when next.js tries to set cookies in a RSC or other enviroments where cookies() is read-only.
Currently, setting a cookie in a read-only cookie enviroment fails silently, which is horrible. The Client is out of sync of the database and you have no idea why.

I had this problem and wasted 6 days working on it. It would save me all this time if I just had this error handling.