issue-2431: feat - fixed prisma schema to avoid while running bunx @better-auth/cli@latest generate
When we run bunx @better-auth/cli@latest generate, prisma will follow
it's naming convention and will name Session[] -> sessions and not
Session, the schema already had two such params Session (as Session[])
and Account (as Account[]), there by duplicate sessions and accounts,
causing issue. Removed the existing Session and Account with sessions
and accounts, hence averting the issue. Furthermore in SsoProvider model
userId was optional but User (which contains userId) was a mandate,
that's not possible. So made User also as optional