fix: Use jwt plugin and RSA instead of HS256 for oidcProvider
Current implementation for jwks in oidcProvider is incomplete / broken
it ignores jwt plugin and uses HS256, which is not normally a
supported algorithm for jwks.

Since changing it to the right algorithm would break any existing
implementations, I added a flag to oidcProvider for "useJWTPlugin"
if that flag is set to true, and jwt plugin is setup, oidcProvider
will use the same key as the one reported in jwks endpoint from jwt
generating tokens that correctly validate.