fix(oauth): Google prompt doesn't allow `+`
In order to get `consent` and `select_account`, we need to pass it like "select_account consent". However BA's type definitions only allow for "select_account+consent" - likely because it's the standard?

Anyway this PR will take that `select_account+consent` defined by the user and convert it to the spaced version (`select_account consent`)