feat: Steam social provider (v2)
Better-Auth does not currently support Steam as a provider because Steam authentication does not comply with OAuth or OIDC standards. For example, it does not respect state, does not require a client ID, client secret, or application registration with redirect whitelisting, and lacks an OIDC endpoint—instead, you must use your own API key to call ISteamUser/GetPlayerSummaries/v002/ after receiving the callback. Additionally, Steam does not return an email address, as user information is only accessible via a public profile endpoint. Due to these incompatibilities, integrating Steam with a spec-compliant adapter is not feasible, so a custom plugin is required.

Big thanks to the initial implementation from https://gist.github.com/Whats-A-MattR/5bce5574e568e8d8e6be55cf692df3a1 while a lot of the logic is re-written, this was a good head start for me!

## Status: ready

- [x] docs