fix: sign-up-email provided dates will cause DB insert failure
When creating an additional field with a type of `date`, then using the `authClient.signUp.email` with that new date field, the endpoint will fail saying it couldn't insert a user.
The error will be `value.toIsoString` is not a function.

This is caused by the client sending dates to the server as a string (since it can't send Date type over HTTP).
Normally, our endpoints which have Zod validation will automatically convert dates from string back to dates, however since the signUpEmail endpoint doesn't do this it causes bugs when inserting.

This PR fixes this and adds tests.