fix(passkey): passkey breaks with `throw: true`
When `throw: true` is configured and is using the passkey plugin, then the result type is unexpected and thus breaks the logic in our client plugin.
This PR addresses this by forcefully setting throw to false, then providing the data as usual, and if there is an error it will handle throwing the error by itself - rather than relying on the authClient.

issue: https://github.com/better-auth/better-auth/issues/2542
linear: https://linear.app/better-auth/issue/ENG-129/fix-passkey-client-returns-options-when-fetch-throws-with-throw-true