fix: logging in `api\index.ts`
It is equivalent now to what it was before. I left a comment why I wanted to change it and maybe anyone can spot it and decide to do that later if I was right. I realized I might have overinterpreted things and it is actually needed to be done like that here. Better safe than sorry