fix: extend default options testing
Removes duplicate/redundant tests and executes them only against relevant option changes. This means some of the previous tests needed to be tested against cache and encryption, turning one test case into 4. Extends default testing to include all the tests that were for the custom options. This resulted in an elegant merge and made the code easier to analyze as well