fix: adapter's factory `debugLog` ignores `options`
Because it has its own debugging config.

`"You are trying to create a record with an id. This is not allowed as we handle id generation for you, unless you pass in the \`forceAllowId\` parameter. The id will be ignored."`
This probably should be printed always as well.

Shouldn't `memory-adapter.ts` use `debugLog` instead of the global logger?

Co-Authored-By: Maxwell <145994855+ping-maxwell@users.noreply.github.com>