fix(magic-link): dont trigger session refresh on sign-in
the magic-link sign-in endpoint doesn't actually sign you in, it's there to send an email for the end user to sign-in via the URL contained in said email.

we automatically trigger a session refresh when the path starts with `/sign-in`, however in this case since the session doesn't actually refresh it doesn't make sense to trigger all useSession hooks to go into `isPending`.