feat: support custom field attributes in plugin schemas
In the auth config, you can define a user/session/account/etc schema to customize the field attributes in a given table.

Right now with plugin schemas, you can only customize the field name. With this PR, will allow support for customizing any field attribute of the field, not just field name.

For example, in the Username plugin, you can't enforce the `username` to be required. With this PR you can do so:
```ts
const auth = betterAuth({
    plugins: [
        username({
            schema: {
                user: {
                    fields: {
                        username: {
                            required: true
                        }
                    }
                }
            }
        })
    ]
})
```