fix(adapter): drizzle `deleteMany` result should be a number
For Neon drizzle driver, the result is an object with other data.
We expect the result of `deleteMany` to be a number representing the number of rows deleted.
This PR fixes this specific bug with neon driver, as well as having additional code to catch unexpected cases with non-numeric return values to warn in the console for them to open an issue as that's not expected.