chore: session refetch cooldown for window focus
Everytime you focus the browser, we make a get-session request.
If you alt+tab between the browser frequently, you would hit your server several times (and thus the DB too) within the span of a couple seconds.
It's unnessesary to get the session each browser refocus within the span of a couple seconds.