fix(multi-session): listDeviceSession and other endpoints breaks with invalid past signatures
Say you use a BETTER_AUTH_SECRET of `1` and signed in, then the next day you changed it to `2`, this would cause the multi-session plugin to crash in various places.

This is because `ctx.getSignedCookie` returns `false` when a cookie has an invalid signature, however due to Better-Call having the incorrect types it caused us to develop Multi-Session incorrectly without considering the possibility that `getSignedCookie` can return false, thus breaking the app.

Better-Call fix to the type issue:
https://github.com/Bekacru/better-call/pull/81