fix(logger): log level priority
Ensure higher severity logs are published when the configured log level is lower (as it should be the "minimum" level to publish).

e.g., `'error'` level messages should be published when BA is configured with log level `'info'`

Unfortunately, this could be considered to be a breaking change. Although it may be what users were expecting all along and could be considered a bug just as much.

Fixes #6408