refactor(kysely): Improve mysql withReturns
When using kysely + mysql then inserting data through the adapter, you may not receive data as the result of the insert. This is because MySQL does not support `RETURNING` in SQL, and thus it does not return any data that was inserted.

It's most troubling when we do not have access to inserting `id`s programmatically, this can be the case when users use `generateId: false` or other auth config options like `generateId: "uuid"` and possibily others.

This forces us to use different query strategies to attempt to get the last-inserted data, however there are still cases where we either can't get that data, or it's unreliable due to something like race-conditions.

This PR introduces a new vector to searching for fields to reliably use to fetch that inserted data, by finding all unique fields in a given schema for a model, we can then see if the provided insert values use data for those unique fields, and use that as a reference to find the inserted data.

Increase withReturns possibility for mysql.