fix(anonymous): missing path breaks anonymous hooks
Recently, we supported `createAuthEndpoint` to not require a `path`, however we forgot to update a specific type which caused the anonmous plugin to have breaking code.

Any users who use the anonymous plugin and accessed an endpoint with not path (such as `auth.api.setPassword`) would run into an error.