fix: BETTER_AUTH_SECRET length check should not throw
In this MR: https://github.com/better-auth/better-auth/pull/6114 there is a check added that throws if the secret is <32 chars.

However for implementation currently running with a secret of <32 this now breaks the application.

I could also not find straight forward migration steps on how to rotate your secret.